requirements.txt
asgiref==3.2.7
click==7.1.1
Django==3.0.5
django-cors-headers==3.3.0
django-rest-auth==0.9.5
django-rq==2.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
EditorConfig==0.12.2
jsbeautifier==1.11.0
PyJWT==1.7.1
pytz==2019.3
redis==3.4.1
rq==1.3.0
six==1.14.0
sqlparse==0.3.1
